Stworzenie klasy Controls i usunięcie ChooseCassette
