--- Assets/dev/wojtek/scripts/MonoBehaviour/ChooseCassette1.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 38f8a836be1053d4f94959c38d281960
